requirements.txt
keras
argparse
tensorflow
Pillow
matplotlib
pydot
numpy
psutil
commentjson
gitpython
colorama
progressbar2